// examples/typescript-api-client/api-client/models/PostBeefRequest.ts
export interface PostBeefRequest {
    /**
     * What state the location is in. Note: Western Australia is split up into two regions, `wa_nw` is North-West Western Australia, `wa_sw` is South-West Western Australia
     * @type {string}
     * @memberof PostBeefRequest
     */
    state: PostBeefRequestStateEnum;
    /**
     * Is this farm north of the Tropic of Capricorn. Note: this is currently approximately -23.43621 degrees latitude
     * @type {boolean}
     * @memberof PostBeefRequest
     */
    northOfTropicOfCapricorn: boolean;
    /**
     * Is there enough rainfall to drain through the soil profile. Note: this is typically above 600mm
     * @type {boolean}
     * @memberof PostBeefRequest
     */
    rainfallAbove600: boolean;
    /**
     * 
     * @type {Array<PostBeefRequestBeefInner>}
     * @memberof PostBeefRequest
     */
    beef: Array<PostBeefRequestBeefInner>;
    /**
     * 
     * @type {Array<PostBeefRequestBurningInner>}
     * @memberof PostBeefRequest
     */
    burning: Array<PostBeefRequestBurningInner>;
    /**
     * 
     * @type {Array<PostBeefRequestVegetationInner>}
     * @memberof PostBeefRequest
     */
    vegetation: Array<PostBeefRequestVegetationInner>;
}


/**
 * @export
 */
export const PostBeefRequestStateEnum = {
    Nsw: 'nsw',
    Vic: 'vic',
    Qld: 'qld',
    Sa: 'sa',
    WaNw: 'wa_nw',
    WaSw: 'wa_sw',
    Tas: 'tas',
    Nt: 'nt',
    Act: 'act'
} as const;
export type PostBeefRequestStateEnum = typeof PostBeefRequestStateEnum[keyof typeof PostBeefRequestStateEnum];