// examples/typescript-api-client/api-client/models/PostBeefRequestBeefInner.ts
export interface PostBeefRequestBeefInner {
    /**
     * Unique identifier for this Beef activity
     * @type {string}
     * @memberof PostBeefRequestBeefInner
     */
    id?: string;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClasses}
     * @memberof PostBeefRequestBeefInner
     */
    classes: PostBeefRequestBeefInnerClasses;
    /**
     * Lime applied in tonnes
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    limestone: number;
    /**
     * Fraction of lime as limestone vs dolomite, between 0 and 1
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    limestoneFraction: number;
    /**
     * 
     * @type {PostBeefRequestBeefInnerFertiliser}
     * @memberof PostBeefRequestBeefInner
     */
    fertiliser: PostBeefRequestBeefInnerFertiliser;
    /**
     * Diesel usage in L (litres)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    diesel: number;
    /**
     * Petrol usage in L (litres)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    petrol: number;
    /**
     * LPG Fuel usage in L (litres)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    lpg: number;
    /**
     * 
     * @type {PostBeefRequestBeefInnerMineralSupplementation}
     * @memberof PostBeefRequestBeefInner
     */
    mineralSupplementation: PostBeefRequestBeefInnerMineralSupplementation;
    /**
     * Source of electricity
     * @type {string}
     * @memberof PostBeefRequestBeefInner
     */
    electricitySource: PostBeefRequestBeefInnerElectricitySourceEnum;
    /**
     * Percent of total electricity usage that is drawn from renewable sources, between 0 and 1. Unused if `electricitySource` is `Renewable`
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    electricityRenewable: number;
    /**
     * Electricity use in KWh (kilowatt hours)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    electricityUse: number;
    /**
     * Grain purchased for cattle feed in tonnes
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    grainFeed: number;
    /**
     * Hay purchased for cattle feed in tonnes
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    hayFeed: number;
    /**
     * Cotton seed purchased for cattle feed in tonnes
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    cottonseedFeed: number;
    /**
     * Total amount of active ingredients of from herbicide (Paraquat, Diquat, Glyphosate) in kg (kilograms)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    herbicide: number;
    /**
     * Total amount of active ingredients of from other herbicides in kg (kilograms)
     * @type {number}
     * @memberof PostBeefRequestBeefInner
     */
    herbicideOther: number;
    /**
     * 
     * @type {PostBeefRequestBeefInnerCowsCalving}
     * @memberof PostBeefRequestBeefInner
     */
    cowsCalving: PostBeefRequestBeefInnerCowsCalving;
}


/**
 * @export
 */
export const PostBeefRequestBeefInnerElectricitySourceEnum = {
    StateGrid: 'State Grid',
    Renewable: 'Renewable'
} as const;
export type PostBeefRequestBeefInnerElectricitySourceEnum = typeof PostBeefRequestBeefInnerElectricitySourceEnum[keyof typeof PostBeefRequestBeefInnerElectricitySourceEnum];