// examples/typescript-api-client/api-client/models/PostBeefRequestBeefInnerClasses.ts
export interface PostBeefRequestBeefInnerClasses {
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    bullsGt1?: PostBeefRequestBeefInnerClassesBullsGt1;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    bullsGt1Traded?: PostBeefRequestBeefInnerClassesBullsGt1Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteersLt1}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steersLt1?: PostBeefRequestBeefInnerClassesSteersLt1;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteersLt1Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steersLt1Traded?: PostBeefRequestBeefInnerClassesSteersLt1Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteers1To2}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steers1To2?: PostBeefRequestBeefInnerClassesSteers1To2;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteers1To2Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steers1To2Traded?: PostBeefRequestBeefInnerClassesSteers1To2Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteersGt2}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steersGt2?: PostBeefRequestBeefInnerClassesSteersGt2;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesSteersGt2Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    steersGt2Traded?: PostBeefRequestBeefInnerClassesSteersGt2Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesCowsGt2}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    cowsGt2?: PostBeefRequestBeefInnerClassesCowsGt2;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesCowsGt2Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    cowsGt2Traded?: PostBeefRequestBeefInnerClassesCowsGt2Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifersLt1}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifersLt1?: PostBeefRequestBeefInnerClassesHeifersLt1;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifersLt1Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifersLt1Traded?: PostBeefRequestBeefInnerClassesHeifersLt1Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifers1To2}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifers1To2?: PostBeefRequestBeefInnerClassesHeifers1To2;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifers1To2Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifers1To2Traded?: PostBeefRequestBeefInnerClassesHeifers1To2Traded;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifersGt2}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifersGt2?: PostBeefRequestBeefInnerClassesHeifersGt2;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesHeifersGt2Traded}
     * @memberof PostBeefRequestBeefInnerClasses
     */
    heifersGt2Traded?: PostBeefRequestBeefInnerClassesHeifersGt2Traded;
}