// examples/typescript-api-client/api-client/models/PostBeefRequestBeefInnerClassesBullsGt1Autumn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AIA Calculator API
 * Emissions Calculators for various farming activities
 *
 * The version of the OpenAPI document: 3.0.0
 * Contact: contact@aginnovationaustralia.com.au
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostBeefRequestBeefInnerClassesBullsGt1Autumn
 */
export interface PostBeefRequestBeefInnerClassesBullsGt1Autumn {
    /**
     * Number of animals (head)
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesBullsGt1Autumn
     */
    head: number;
    /**
     * Average liveweight of animals in kg/head (kilogram per head)
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesBullsGt1Autumn
     */
    liveweight: number;
    /**
     * Average liveweight gain in kg/day (kilogram per day)
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesBullsGt1Autumn
     */
    liveweightGain: number;
    /**
     * Crude protein percent, between 0 and 100
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesBullsGt1Autumn
     */
    crudeProtein?: number;
    /**
     * Dry matter digestibility percent, between 0 and 100
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesBullsGt1Autumn
     */
    dryMatterDigestibility?: number;
}

/**
 * Check if a given object implements the PostBeefRequestBeefInnerClassesBullsGt1Autumn interface.
 */
export function instanceOfPostBeefRequestBeefInnerClassesBullsGt1Autumn(value: object): value is PostBeefRequestBeefInnerClassesBullsGt1Autumn {
    if (!('head' in value) || value['head'] === undefined) return false;
    if (!('liveweight' in value) || value['liveweight'] === undefined) return false;
    if (!('liveweightGain' in value) || value['liveweightGain'] === undefined) return false;
    return true;
}

export function PostBeefRequestBeefInnerClassesBullsGt1AutumnFromJSON(json: any): PostBeefRequestBeefInnerClassesBullsGt1Autumn {
    return PostBeefRequestBeefInnerClassesBullsGt1AutumnFromJSONTyped(json, false);
}

export function PostBeefRequestBeefInnerClassesBullsGt1AutumnFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostBeefRequestBeefInnerClassesBullsGt1Autumn {
    if (json == null) {
        return json;
    }
    return {
        
        'head': json['head'],
        'liveweight': json['liveweight'],
        'liveweightGain': json['liveweightGain'],
        'crudeProtein': json['crudeProtein'] == null ? undefined : json['crudeProtein'],
        'dryMatterDigestibility': json['dryMatterDigestibility'] == null ? undefined : json['dryMatterDigestibility'],
    };
}

export function PostBeefRequestBeefInnerClassesBullsGt1AutumnToJSON(json: any): PostBeefRequestBeefInnerClassesBullsGt1Autumn {
    return PostBeefRequestBeefInnerClassesBullsGt1AutumnToJSONTyped(json, false);
}

export function PostBeefRequestBeefInnerClassesBullsGt1AutumnToJSONTyped(value?: PostBeefRequestBeefInnerClassesBullsGt1Autumn | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'head': value['head'],
        'liveweight': value['liveweight'],
        'liveweightGain': value['liveweightGain'],
        'crudeProtein': value['crudeProtein'],
        'dryMatterDigestibility': value['dryMatterDigestibility'],
    };
}