// examples/typescript-api-client/api-client/models/PostBeefRequestBeefInnerClassesCowsGt2.ts
export interface PostBeefRequestBeefInnerClassesCowsGt2 {
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1Autumn}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    autumn: PostBeefRequestBeefInnerClassesBullsGt1Autumn;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1Autumn}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    winter: PostBeefRequestBeefInnerClassesBullsGt1Autumn;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1Autumn}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    spring: PostBeefRequestBeefInnerClassesBullsGt1Autumn;
    /**
     * 
     * @type {PostBeefRequestBeefInnerClassesBullsGt1Autumn}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    summer: PostBeefRequestBeefInnerClassesBullsGt1Autumn;
    /**
     * Number of animals sold (head)
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    headSold: number;
    /**
     * Weight at sale, in liveweight kg/head (kilogram per head)
     * @type {number}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    saleWeight: number;
    /**
     * 
     * @type {Array<PostBeefRequestBeefInnerClassesBullsGt1PurchasesInner>}
     * @memberof PostBeefRequestBeefInnerClassesCowsGt2
     */
    purchases?: Array<PostBeefRequestBeefInnerClassesBullsGt1PurchasesInner>;
}