// examples/typescript-api-client/api-client/models/PostBeefRequestBeefInnerCowsCalving.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AIA Calculator API
 * Emissions Calculators for various farming activities
 *
 * The version of the OpenAPI document: 3.0.0
 * Contact: contact@aginnovationaustralia.com.au
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Fraction of cows calving in each season, between 0 and 1
 * @export
 * @interface PostBeefRequestBeefInnerCowsCalving
 */
export interface PostBeefRequestBeefInnerCowsCalving {
    /**
     * 
     * @type {number}
     * @memberof PostBeefRequestBeefInnerCowsCalving
     */
    spring: number;
    /**
     * 
     * @type {number}
     * @memberof PostBeefRequestBeefInnerCowsCalving
     */
    summer: number;
    /**
     * 
     * @type {number}
     * @memberof PostBeefRequestBeefInnerCowsCalving
     */
    autumn: number;
    /**
     * 
     * @type {number}
     * @memberof PostBeefRequestBeefInnerCowsCalving
     */
    winter: number;
}

/**
 * Check if a given object implements the PostBeefRequestBeefInnerCowsCalving interface.
 */
export function instanceOfPostBeefRequestBeefInnerCowsCalving(value: object): value is PostBeefRequestBeefInnerCowsCalving {
    if (!('spring' in value) || value['spring'] === undefined) return false;
    if (!('summer' in value) || value['summer'] === undefined) return false;
    if (!('autumn' in value) || value['autumn'] === undefined) return false;
    if (!('winter' in value) || value['winter'] === undefined) return false;
    return true;
}

export function PostBeefRequestBeefInnerCowsCalvingFromJSON(json: any): PostBeefRequestBeefInnerCowsCalving {
    return PostBeefRequestBeefInnerCowsCalvingFromJSONTyped(json, false);
}

export function PostBeefRequestBeefInnerCowsCalvingFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostBeefRequestBeefInnerCowsCalving {
    if (json == null) {
        return json;
    }
    return {
        
        'spring': json['spring'],
        'summer': json['summer'],
        'autumn': json['autumn'],
        'winter': json['winter'],
    };
}

export function PostBeefRequestBeefInnerCowsCalvingToJSON(json: any): PostBeefRequestBeefInnerCowsCalving {
    return PostBeefRequestBeefInnerCowsCalvingToJSONTyped(json, false);
}

export function PostBeefRequestBeefInnerCowsCalvingToJSONTyped(value?: PostBeefRequestBeefInnerCowsCalving | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'spring': value['spring'],
        'summer': value['summer'],
        'autumn': value['autumn'],
        'winter': value['winter'],
    };
}